FEARING SYMPATHY Matrix
This is a Yi Jing based matrix meant to transcribe base64 information into the form of plot elements.  Specific words or characters need not be used -- the goal is to encode a link in the form of *ideas*.  This is cumbersome, but a philosophically interesting exercise.  For example, the original New York Post tinyurl link is y5zr43pm = Crisis,Honesty,Politeness,Alienation,The underground,Discretion,Punishment,The problem solver.  The link might be coded (sorry for the lack of inspired prose) as "Global warming is a problem.  Coal miners are good upstanding people.  They should be addressed with respect.  Some people just get disgusted with the climate situation and don't care about anything.  There is a network of opposition where people trade crazy ideas.  The government doesn't know how to handle them and is holding off action.  But there are things they could be charged with.  But there are methods for carbon sequestration that will set everything straight."  It may be a bit of a loopy comment ... but what machine will recognize it as a WEB LINK?  It will be interesting to see whether the accuracy of translation can be adequate.