Add files via upload
This is quite a shaggy dog of a file, but it includes a line by line comparison of all the Yi Jing hexagrams as they were altered somewhat to fit a more mundane set of meanings.  This might help with translating to and from the hexagram format somewhat, but mostly this is included out of some interest in Yi Jing itself.